Extract protected_request into a module
We want to protect even more requests (ie. in the Source class).
