+++ lib/solidus_paypal_braintree.rb
@@ -2,6 +2,7 @@ require 'solidus_core'
 require 'solidus_support'
 require 'solidus_paypal_braintree/engine'
 require 'solidus_paypal_braintree/country_mapper'
+require 'solidus_paypal_braintree/request_protection'
 
 module SolidusPaypalBraintree
   def self.table_name_prefix
